+++ src/components/ui/PageTitle.tsx
@@ -1,4 +1,4 @@
-import Typography from "@mui/material/Typography";
+import Typography, { TypographyProps } from "@mui/material/Typography";
 
 import Span from "@/components/layout/Span";
 import useDocumentTitle from "@/hooks/useDocumentTitle";
@@ -6,13 +6,14 @@ import useDocumentTitle from "@/hooks/useDocumentTitle";
 type Props = {
   value: string;
   mobileValue?: string | undefined;
+  mb?: TypographyProps["mb"] | undefined;
 };
 
-const PageTitle = ({ value, mobileValue }: Props): React.ReactNode => {
+const PageTitle = ({ value, mobileValue, mb = 2 }: Props): React.ReactNode => {
   useDocumentTitle(value);
 
   return (
-    <Typography variant="h2" className="ycc-page-title" mb={2}>
+    <Typography variant="h2" className="ycc-page-title" mb={mb}>
       <Span sx={{ display: { xs: "inline", sm: "none" } }}>
         {mobileValue ?? value}
       </Span>