+++ src/context/shared-data/useLicenceInfos.ts
@@ -3,7 +3,7 @@ import { LicenceDetailedInfo } from "@/model/dtos";
 
 import useSharedData from "./useSharedData";
 
-const useLicenceInfos = (): PromiseOutcome<Readonly<LicenceDetailedInfo[]>> => {
+const useLicenceInfos = (): PromiseOutcome<readonly LicenceDetailedInfo[]> => {
   const sharedData = useSharedData();
   return usePromise(sharedData.getLicenceInfos, []);
 };