+++ src/pages/helpers/task/HelperTaskForm.tsx
@@ -45,15 +45,19 @@ import {
 } from "@/pages/helpers/helpers-utils";
 import { getFullNameAndUsername } from "@/pages/members/members-utils";
 import client from "@/utils/client";
+import {
+  HELPER_TASK_SHORT_DESCRIPTION_MAX_LENGTH,
+  HELPER_TASK_TITLE_MAX_LENGTH,
+} from "@/utils/constants";
 import { getNow } from "@/utils/date-utils";
 import dayjs from "@/utils/dayjs";
 
 type Props = {
   task: HelperTask | undefined;
   newTask: boolean;
-  categories: Readonly<HelperTaskCategory[]>;
-  members: Readonly<MemberPublicInfo[]>;
-  licenceInfos: Readonly<LicenceDetailedInfo[]>;
+  categories: readonly HelperTaskCategory[];
+  members: readonly MemberPublicInfo[];
+  licenceInfos: readonly LicenceDetailedInfo[];
 };
 
 type HelperTaskCreationRequestExtra = Omit<
@@ -86,9 +90,7 @@ const HelperTaskForm = ({
 }: Props): React.ReactNode => {
   const currentUser = useCurrentUser();
   const [error, setError] = useState<unknown>();
-  const longDescription = useDelayedRef<string | null | undefined>(
-    task?.longDescription,
-  );
+  const longDescription = useDelayedRef(task?.longDescription);
   const [type, setType] = useState(task?.type ?? HelperTaskType.Shift);
   const [multiDayShift, setMultiDayShift] = useState(
     task ? isMultiDayShift(task) : false,
@@ -308,7 +310,9 @@ const HelperTaskForm = ({
         title: "Please confirm the following:",
         content,
         delayConfirm: true,
-        onConfirm: async () => await doSubmit(base, creation, update),
+        onConfirm: async () => {
+          await doSubmit(base, creation, update);
+        },
       });
     } else {
       await doSubmit(base, creation, update);
@@ -375,7 +379,15 @@ const HelperTaskForm = ({
       </SpacedBox>
 
       <SpacedBox>
-        <TextFieldElement name="base.title" required label="Title" fullWidth />
+        <TextFieldElement
+          name="base.title"
+          required
+          label="Title"
+          fullWidth
+          slotProps={{
+            htmlInput: { maxLength: HELPER_TASK_TITLE_MAX_LENGTH },
+          }}
+        />
       </SpacedBox>
 
       <SpacedBox>
@@ -384,6 +396,11 @@ const HelperTaskForm = ({
           required
           label="Short Description (included in emails)"
           fullWidth
+          slotProps={{
+            htmlInput: {
+              maxLength: HELPER_TASK_SHORT_DESCRIPTION_MAX_LENGTH,
+            },
+          }}
         />
       </SpacedBox>
 
@@ -486,7 +503,9 @@ const HelperTaskForm = ({
                 control={
                   <Switch
                     checked={multiDayShift}
-                    onChange={(_, checked) => setMultiDayShift(checked)}
+                    onChange={(_, checked) => {
+                      setMultiDayShift(checked);
+                    }}
                   />
                 }
                 label="Multi-day shift"